requirements.txt
django
django-simple-sso
django-session-timeout